build fix
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@64858 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Runtime/tools/RTSignTool.cpp
@@ -662,7 +662,7 @@ static int HandleShowExeWorkerPkcs7Decode(PSHOWEXEPKCS7 pThis)
 {
     RTERRINFOSTATIC     ErrInfo;
     RTASN1CURSORPRIMARY PrimaryCursor;
-    RTAsn1CursorInitPrimary(&PrimaryCursor, pThis->pbBuf, pThis->cbBuf, RTErrInfoInitStatic(&ErrInfo),
+    RTAsn1CursorInitPrimary(&PrimaryCursor, pThis->pbBuf, (uint32_t)pThis->cbBuf, RTErrInfoInitStatic(&ErrInfo),
                             &g_RTAsn1DefaultAllocator, 0, "WinCert");
 
     int rc = RTCrPkcs7ContentInfo_DecodeAsn1(&PrimaryCursor.Cursor, 0, &pThis->ContentInfo, "CI");
